meu-servidor-ia/requirements/voice.txt
# Voice Generation Service Dependencies
-r base.txt

# Audio Processing
torchaudio>=2.1.1
soundfile>=0.12.1
librosa>=0.10.1
scipy>=1.11.4
praat-parselmouth>=0.4.3

# Fish Speech v2
fairseq>=0.12.2
gradio>=4.7.1
ffmpeg-python>=0.2.0
pydub>=0.25.1 